test renaming of norm_fiber to norm_fiber_inplace
