+++ include/CMakeLists.txt
@@ -75,7 +75,7 @@ set(KERNEL_HDR
     "nntile/kernel/sum_slice/cpu.hh"
     "nntile/kernel/sum_fiber.hh"
     "nntile/kernel/sum_fiber/cpu.hh"
-    "nntile/kernel/norm_fiber/cpu.hh"
+    "nntile/kernel/norm_fiber_inplace/cpu.hh"
     "nntile/kernel/norm_slice.hh"
     "nntile/kernel/norm_slice/cpu.hh"
     "nntile/kernel/pow.hh"
@@ -172,7 +172,7 @@ if(NNTILE_USE_CUDA)
         "nntile/kernel/fill/cuda.hh"
         "nntile/kernel/sum_slice/cuda.hh"
         "nntile/kernel/sum_fiber/cuda.hh"
-        "nntile/kernel/norm_fiber/cuda.hh"
+        "nntile/kernel/norm_fiber_inplace/cuda.hh"
         "nntile/kernel/logsumexp/cuda.hh"
         "nntile/kernel/norm_slice/cuda.hh"
         "nntile/kernel/pow/cuda.hh"
