+++ include/nntile/kernel/norm_fiber_inplace.hh
@@ -6,7 +6,7 @@
  * NNTile is software framework for fast training of big neural networks on
  * distributed-memory heterogeneous systems based on StarPU runtime system.
  *
- * @file include/nntile/kernel/norm_fiber.hh
+ * @file include/nntile/kernel/norm_fiber_inplace.hh
  * Euclidean norms over slices into a fiber of a product of a buffer
  *
  * @version 1.1.0
@@ -14,16 +14,16 @@
 
 #pragma once
 
-#include <nntile/kernel/norm_fiber/cpu.hh>
+#include <nntile/kernel/norm_fiber_inplace/cpu.hh>
 #include <nntile/defs.h>
 #ifdef NNTILE_USE_CUDA
-#include <nntile/kernel/norm_fiber/cuda.hh>
+#include <nntile/kernel/norm_fiber_inplace/cuda.hh>
 #endif // NNTILE_USE_CUDA
 
-//! @namespace nntile::kernel::norm_fiber
-/*! Low-level implementations of computing norm_fiber operation
+//! @namespace nntile::kernel::norm_fiber_inplace
+/*! Low-level implementations of computing norm_fiber_inplace operation
  * */
-namespace nntile::kernel::norm_fiber
+namespace nntile::kernel::norm_fiber_inplace
 {
 
-} // namespace nntile::kernel::norm_fiber
+} // namespace nntile::kernel::norm_fiber_inplace
