+++ include/nntile/kernel/norm_fiber_inplace/cpu.hh
@@ -6,7 +6,7 @@
  * NNTile is software framework for fast training of big neural networks on
  * distributed-memory heterogeneous systems based on StarPU runtime system.
  *
- * @file include/nntile/kernel/norm_fiber/cpu.hh
+ * @file include/nntile/kernel/norm_fiber_inplace/cpu.hh
  * Euclidean norms over slices into a fiber of a product of buffers on CPU
  *
  * @version 1.1.0
@@ -16,7 +16,7 @@
 
 #include <nntile/base_types.hh>
 
-namespace nntile::kernel::norm_fiber
+namespace nntile::kernel::norm_fiber_inplace
 {
 
 // Euclidean norms over slices into a fiber of a product of buffers on CPU
@@ -24,4 +24,4 @@ template<typename T>
 void cpu(Index m, Index n, Index k, Index batch, Scalar alpha_, const T *src, Scalar beta_, T *dst)
     noexcept;
 
-} // namespace nntile::kernel::norm_fiber
+} // namespace nntile::kernel::norm_fiber_inplace
